changed fd type from size_t to int

+++ ft_putstri_fd.c
@@ -6,13 +6,13 @@
 /*   By: sbonnefo <sbonnefo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/25 15:45:19 by sbonnefo          #+#    #+#             */
-/*   Updated: 2017/01/25 15:45:49 by sbonnefo         ###   ########.fr       */
+/*   Updated: 2017/01/25 15:49:40 by sbonnefo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		ft_putstri_fd(char const *s, size_t len, size_t fd)
+void		ft_putstri_fd(char const *s, size_t len, int fd)
 {
 	write(fd, s, len);
 }
